+++ NEWS.md
@@ -1,3 +1,9 @@
+# scicomptools 1.1.0
+
+- New function: `molec_wt`. Identifies molecular weight for specified element based on element name, symbol, or atomic number.
+- New function: `issue_extract`. Exports specified GitHub issues as PDF files when given the URL of a GitHub repository and a numeric vector of GitHub issue numbers. 
+- Unit tests added for existing functions in the package. Likely limited impact on users except better adherence to best practice will make for a stronger package in the long run.
+
 # scicomptools 1.0.0
 
 This is the first fully functioning version of the package. There are currently no ERRORs, WARNINGs, or NOTEs returned by `devtools::check()`